+++ backend/app/router/v1/__init__.py
@@ -1,0 +1,1 @@
+from app.router.v1.router import *
